requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
chardet==3.0.4
discord.py==1.4.1
dnspython==2.0.0
idna==2.10
mongoengine==0.20.0
multidict==4.7.6
pycodestyle==2.6.0
pymongo==3.11.0
python-dotenv==0.14.0
toml==0.10.1
yarl==1.5.1
